Add files via upload
The significance of this android app that uses a proximity sensor and screen orientation sensor is that it allows users to track and monitor their press ups workout progress. The login and registration feature allow users to save and track their progress over time, and potentially compete with other users or share their progress on social media. the app utilizes MySQL database for data storage and retrieval. Additionally, the use of sensors provides more accurate tracking of press-ups, as it can detect when a user is properly lowering and raising their body during the exercise. This helps users ensure that they are performing the exercise correctly and effectively.